+++ src/qanGraph.h
@@ -157,7 +157,8 @@ public:
     Q_INVOKABLE void    setConnectorSource(qan::Node* sourceNode) noexcept;
 signals:
     //! \copydoc hlg::Connector::requestEdgeCreation
-    void                connectorRequestEdgeCreation(qan::Node* src, QObject* dst);
+    void                connectorRequestEdgeCreation(qan::Node* src, QObject* dst,
+                                                     qan::PortItem* srcPortItem, qan::PortItem* dstPortItem);
     //! \copydoc hlg::Connector::edgeInserted
     void                connectorEdgeInserted(qan::Edge* edge);
 
@@ -294,7 +295,7 @@ public: // should be considered private
      * \arg parent Returned selection item is automatically reparented to \c parent (could be nullptr).
      * \return A selection item or nullptr if graph \c selectionDelegate is invalid, ownershipd goes to the caller with QmlEngine::CppOwnership, might be nullptr.
      */
-    QPointer<QQuickItem>    createSelectionItem(QQuickItem* parent) noexcept;
+    QQuickItem*             createSelectionItem(QQuickItem* parent);
 protected:
 
     struct QObjectDeleteLater {
@@ -308,9 +309,9 @@ protected:
     std::unique_ptr<QQmlComponent>  _selectionDelegate{nullptr};
 private:
     //! Secure factory for QML components, errors are reported on stderr.
-    std::unique_ptr<QQmlComponent>  createComponent(const QString& url) noexcept;
+    std::unique_ptr<QQmlComponent>  createComponent(const QString& url);
     //! Secure utility to create a QQuickItem from a given QML component \c component (might issue warning if component is nullptr or not successfully loaded).
-    QPointer<QQuickItem>            createItemFromComponent(QQmlComponent* component) noexcept;
+    QQuickItem*                     createItemFromComponent(QQmlComponent* component);
     //@}
     //-------------------------------------------------------------------------
 
@@ -683,11 +684,16 @@ public:
     //! Similar to selectNode() for qan::Group (internally group is a node).
     bool            selectGroup(qan::Group& group, Qt::KeyboardModifiers modifiers = Qt::NoModifier);
 
+    //! Similar to selectNode() for qan::Edge.
+    bool            selectEdge(qan::Edge& edge, Qt::KeyboardModifiers modifiers = Qt::NoModifier);
+
     /*! \brief Add a node in the current selection.
      */
     void            addToSelection(qan::Node& node);
     //! \copydoc addToSelection
-    void            addToSelection(qan::Group& node);
+    void            addToSelection(qan::Group& group);
+    //! \copydoc addToSelection
+    void            addToSelection(qan::Edge& edge);
 
     //! Remove a node from the selection.
     void            removeFromSelection(qan::Node& node);
@@ -715,8 +721,8 @@ public:
     Q_PROPERTY(QAbstractItemModel* selectedNodes READ getSelectedNodesModel NOTIFY selectedNodesChanged FINAL)  // In fact non-notifiable, avoid QML warning
     QAbstractItemModel* getSelectedNodesModel() { return qobject_cast<QAbstractItemModel*>(_selectedNodes.model()); }
 
-    inline auto         getSelectedNodes() noexcept -> SelectedNodes& { return _selectedNodes; }
-    inline auto         getSelectedNodes() const noexcept -> const SelectedNodes& { return _selectedNodes; }
+    inline auto         getSelectedNodes() -> SelectedNodes& { return _selectedNodes; }
+    inline auto         getSelectedNodes() const -> const SelectedNodes& { return _selectedNodes; }
 private:
     SelectedNodes       _selectedNodes;
 signals:
@@ -729,12 +735,26 @@ public:
     Q_PROPERTY(QAbstractItemModel* selectedGroups READ getSelectedGroupsModel NOTIFY selectedGroupsChanged FINAL)   // In fact non-notifiable, avoid QML warning
     QAbstractItemModel* getSelectedGroupsModel() { return qobject_cast<QAbstractItemModel*>(_selectedGroups.model()); }
 
-    inline auto         getSelectedGroups() noexcept -> SelectedGroups& { return _selectedGroups; }
-    inline auto         getSelectedGroups() const noexcept -> const SelectedGroups& { return _selectedGroups; }
+    inline auto         getSelectedGroups() -> SelectedGroups& { return _selectedGroups; }
+    inline auto         getSelectedGroups() const -> const SelectedGroups& { return _selectedGroups; }
 private:
     SelectedGroups      _selectedGroups;
 signals:
     void                selectedGroupsChanged();
+
+public:
+    using SelectedEdges = QVector<QPointer<qan::Edge>>;
+
+    //! Read-only list model of currently selected edges.
+    //Q_PROPERTY(QAbstractItemModel* selectedEdges READ getSelectedEdgesModel NOTIFY selectedEdgesChanged FINAL)   // In fact non-notifiable, avoid QML warning
+    //QAbstractItemModel* getSelectedEdgesModel() { return qobject_cast<QAbstractItemModel*>(_selectedEdges.model()); }
+
+    inline auto         getSelectedEdges() -> SelectedEdges& { return _selectedEdges; }
+    inline auto         getSelectedEdges() const -> const SelectedEdges& { return _selectedEdges; }
+private:
+    SelectedEdges       _selectedEdges;
+signals:
+    void                selectedEdgesChanged();
 
 protected:
     //! \brief Return a vector of currently selected nodes/groups items.